# test-api-server/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Test API Server
  description: |
    A comprehensive API server for testing MCP Studio functionality. 
    Provides CRUD operations for users, chats, messages, and feedbacks with MongoDB backend.

    This API is designed to be easily convertible to an MCP (Model Context Protocol) server.
  version: 1.0.0
  contact:
    name: MCP Studio Team
    email: support@mcpstudio.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.mcpstudio.com
    description: Production server
...
paths:
...
  /api/messages:
    get:
      tags: [messages]
      summary: Get all messages
      description: Retrieve a paginated list of all messages
      operationId: getMessages
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "500":
          description: Failed to fetch messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [messages]
      summary: Create message
      description: Create a new message in a chat
      operationId: createMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: Message created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMessageResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to create message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/messages/{id}:
    get:
      tags: [messages]
      summary: Get message by ID
      description: Retrieve a specific message by its ID
      operationId: getMessageById
      parameters:
        - name: id
          in: path
          required: true
          description: Message ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Message retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Invalid message ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [messages]
      summary: Update message
      description: Update an existing message
      operationId: updateMessage
      parameters:
        - name: id
          in: path
          required: true
          description: Message ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMessageRequest"
      responses:
        "200":
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponse"
        "400":
          description: Invalid input or message ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to update message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [messages]
      summary: Delete message
      description: Delete a message by its ID
      operationId: deleteMessage
      parameters:
        - name: id
          in: path
          required: true
          description: Message ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Message deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "400":
          description: Invalid message ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to delete message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/messages/chat/{chatId}:
    get:
      tags: [messages]
      summary: Get messages by chat ID
      description: Retrieve all messages for a specific chat
      operationId: getMessagesByChatId
      parameters:
        - name: chatId
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - name: limit
          in: query
          description: Number of messages per page (default 50)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "500":
          description: Failed to fetch messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [messages]
      summary: Delete all messages for chat
      description: Delete all messages in a specific chat
      operationId: deleteMessagesByChatId
      parameters:
        - name: chatId
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
      responses:
        "200":
          description: Messages deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "500":
          description: Failed to delete messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Feedback endpoints
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination (optional - only used when both page and limit are provided)
      schema:
        type: integer
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page (optional - only used when both page and limit are provided, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
  schemas:
...
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      description: MongoDB ObjectId
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 timestamp

    # Error responses
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message

    # Pagination (optional - only when page and limit are both provided)
    PaginationOptional:
      type: object
      properties:
        current:
          type: integer
          description: Current page number
        total:
          type: integer
          description: Total number of pages
        count:
          type: integer
          description: Number of items in current page
        totalRecords:
          type: integer
          description: Total number of records

    # Base list response without pagination
...
    Message:
      type: object
      required:
        - _id
        - chatId
        - content
        - createdAt
        - updatedAt
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        chatId:
          type: string
        content:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    CreateMessageRequest:
      type: object
      required:
        - chatId
        - content
      properties:
        chatId:
          type: string
        content:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    UpdateMessageRequest:
      type: object
      properties:
        content:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        metadata:
          type: object
          additionalProperties: true
...
    CreateMessageResponse:
      type: object
      required:
        - message
        - data
      properties:
        message:
          type: string
        data:
          $ref: "#/components/schemas/Message"
...
    MessageListResponse:
      type: object
      required:
        - messages
        - totalRecords
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        totalRecords:
          type: integer
          description: Total number of messages in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"

    # Feedback schemas
...
    UpdateResponse:
      type: object
      required:
        - message
        - modifiedCount
      properties:
        message:
          type: string
        modifiedCount:
          type: integer
    DeleteResponse:
      type: object
      required:
        - message
        - deletedCount
      properties:
        message:
          type: string
        deletedCount:
          type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - bearerAuth: []